Move TransactionResult::fromSegmentedValues to TransactionMessage class
